// FrontEnd/amaris-app/src/app/services/fondos-api/model/backendFondosApplicationDTOsFondoDto.ts
/**
 * BackendFondos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BackendFondosApplicationDTOsFondoDto { 
    fondoID?: string;
    nombreFondo?: string;
    montoMinimo?: number;
    categoria?: string;
}