// FrontEnd/amaris-app/src/app/services/fondos-api/model/backendFondosApplicationDTOsResultadoOperacionDto.ts
/**
 * BackendFondos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TipoTransaccion } from './tipoTransaccion';


export interface BackendFondosApplicationDTOsResultadoOperacionDto { 
    exito?: boolean;
    mensajeNotificacion?: string;
    clienteId?: string;
    fondoId?: string;
    tipo?: TipoTransaccion;
}
export namespace BackendFondosApplicationDTOsResultadoOperacionDto {
}