// FrontEnd/amaris-app/src/app/services/fondos-api/model/backendFondosApplicationDTOsTransaccionDto.ts
/**
 * BackendFondos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BackendFondosApplicationDTOsTransaccionDto { 
    tipo?: string;
    fondoID?: string;
    monto?: number;
    fecha?: string;
    notificacion?: string;
}